UI Bugfix (Zeit/Level wurden fälschlicherweise gerundet)

+++ src/main/java/de/uzk/gui/areas/AreaStatsBar.java
@@ -83,7 +83,7 @@ public class AreaStatsBar extends ObserverContainer<JPanel> {
     }
 
     private void updateLevel() {
-        double level = (int) (workspace.getLevel() * workspace.getConfig().getLevelUnit());
+        double level =  (workspace.getLevel() * workspace.getConfig().getLevelUnit());
         String levelString = "%.01f μm".formatted(level);
         this.labelLevel.setText(StringUtils.wrapHtml(StringUtils.applyFontSize(
             levelString, 75)));
